% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{FUNCTION_TITLE}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{PARAM_DESCRIPTION}

\item{output}{PARAM_DESCRIPTION}

\item{session}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
